Changing height, radioview added

+++ GeoObject.cpp
@@ -1,11 +1,11 @@
 #include "GeoObject.h"
 
-GeoObject::GeoObject(float nx, float ny, float nr, bool npvo) : Object(nx, ny, npvo)
+GeoObject::GeoObject(double nx, double ny, double nr, bool npvo) : Object(nx, ny, npvo)
 {
 	radius = nr;
     type = 0;
 };
-GeoObject::GeoObject(float nx1, float ny1, float nx2, float ny2, float nx3, float ny3, float nx4, float ny4, bool npvo) : Object((nx1 + nx2 + nx3 + nx4)/4, (ny1 + ny2 + ny3 + ny4)/4, npvo)
+GeoObject::GeoObject(double nx1, double ny1, double nx2, double ny2, double nx3, double ny3, double nx4, double ny4, bool npvo) : Object((nx1 + nx2 + nx3 + nx4)/4, (ny1 + ny2 + ny3 + ny4)/4, npvo)
 {
      vector<Object> angle1, angle2;
      angles[0].setX(nx1);
@@ -53,7 +53,7 @@ GeoObject::GeoObject(float nx1, float ny1, float nx2, float ny2, float nx3, floa
      radius = max(radius, distance(angles[2]));
      radius = max(radius, distance(angles[3]));
 };
-bool GeoObject::onWay(float x1, float y1, float x2, float y2)
+bool GeoObject::onWay(double x1, double y1, double x2, double y2)
 {
 	if (x1 == x2)
 	{
@@ -64,16 +64,16 @@ bool GeoObject::onWay(float x1, float y1, float x2, float y2)
 	if ((x >= min(x1, x2) && x <= max(x1, x2) && y >= min(y1, y2) && y <= max(y1, y2)) || fabs(x - x2) <= radius || fabs(y - y2) <= radius)
 	{
 		//попадает в прямоугольник. может, использовать его в начале поиска оптимальной точки, а не снаружи?
-		float k = (y2 - y1)/(x2 - x1);
-		float b = y1 - x1*(y2 - y1)/(x2 - x1);
-		float d = (pow((2*k*b - 2*x - 2*y*k), 2) - (4 + 4*k*k)*(b*b - radius*radius + x*x + y*y - 2*y*b));   
+        double k = (y2 - y1)/(x2 - x1);
+        double b = y1 - x1*(y2 - y1)/(x2 - x1);
+        double d = (pow((2*k*b - 2*x - 2*y*k), 2) - (4 + 4*k*k)*(b*b - radius*radius + x*x + y*y - 2*y*b));
 		if(d < 0) 
 			return false;
 		return true;
 	}
 	return false;
 };
-Object GeoObject::point(float x1, float y1, float x2, float y2)
+Object GeoObject::point(double x1, double y1, double x2, double y2)
 {
 	if (abs(x1 - x2) <= abs(y1 - y2)/100)
 	{
@@ -87,12 +87,12 @@ Object GeoObject::point(float x1, float y1, float x2, float y2)
 			return Object(x, y + radius*6.0/5.0);
 		return Object(x, y - radius*6.0/5.0);
 	}
-	float k = (y2 - y1)/(x2 - x1);
-	float b = y1 - x1*(y2 - y1)/(x2 - x1);
-	float d = pow((2*k*b - 2*x - 2*y*k), 2) - (4 + 4*k*k)*(b*b - radius*radius + x*x + y*y - 2*y*b);
-	float x1cross = ((-(2*k*b - 2*x - 2*y*k) - sqrt(d))/(2 + 2*k*k)), x2cross = ((-(2*k*b - 2*x - 2*y*k) + sqrt(d))/(2 + 2*k*k)); 
-	float y1cross = k*x1cross + b, y2cross = k*x2cross + b;
-	float xcentre = (x1cross + x2cross)/2, ycentre = k*xcentre + b;
+    double k = (y2 - y1)/(x2 - x1);
+    double b = y1 - x1*(y2 - y1)/(x2 - x1);
+    double d = pow((2*k*b - 2*x - 2*y*k), 2) - (4 + 4*k*k)*(b*b - radius*radius + x*x + y*y - 2*y*b);
+    double x1cross = ((-(2*k*b - 2*x - 2*y*k) - sqrt(d))/(2 + 2*k*k)), x2cross = ((-(2*k*b - 2*x - 2*y*k) + sqrt(d))/(2 + 2*k*k));
+    double y1cross = k*x1cross + b, y2cross = k*x2cross + b;
+    double xcentre = (x1cross + x2cross)/2, ycentre = k*xcentre + b;
 	//cout << xcentre << " " << ycentre << '\n';
 	/*if (xcentre == x)
 	{
@@ -118,7 +118,7 @@ bool GeoObject::operator==(GeoObject go2)
 		return true;
 	return false;
 };
-float GeoObject::getRadius()
+double GeoObject::getRadius()
 {
 	return radius;
 };
