+++ PVO.h
@@ -2,8 +2,8 @@
 
 class PVO: public GeoObject //переопределить isPVO как true
 {
-	float p; //вероятность быть сбитым в этой области
+	double p; //вероятность быть сбитым в этой области
 public:
-	PVO(float nx, float ny, float nr, float np); // сначала вызвать конструктор родительского!
-	bool compile(float xuva, float yuva, float zuva); //возвращает true, если координаты попадают в полусферу и БЛА сбит с учетом вероятности
+	PVO(double nx, double ny, double nr, double np); // сначала вызвать конструктор родительского!
+	bool compile(double xuva, double yuva, double zuva); //возвращает true, если координаты попадают в полусферу и БЛА сбит с учетом вероятности
 };
